Refactor: Rename deep_validator.py to code_metrics.py
- Renamed `agent/deep_validator.py` to `agent/code_metrics.py` to better reflect its content focusing on code complexity, duplication, and metric analysis.
- Updated all internal import statements in `agent/__init__.py` and `agent/code_validator.py` to use the new module name.
- Renamed the corresponding test file from `tests/test_deep_validator.py` to `tests/test_code_metrics.py`.
- Updated imports within the test file to reflect the module rename.

Note: Pytest execution was hindered by a persistent sandbox environment issue, but manual code and structure verification indicates changes are correctly applied.
