+++ config/models/main.yaml
@@ -1,0 +1,38 @@
+# Hydra Group Configuration for Models (Located in config/models/main.yaml)
+
+defaults:
+  - /base_config/default_model_settings # Inherit default temperature and max_tokens
+  # This structure assumes base_config.yaml is at the root of the config dir.
+  # The / prefix makes it an absolute path from the config root.
+
+architect_default:
+  primary: "gemini/gemini-2.5-pro"
+  fallback: "deepseek/deepseek-r1-0528:free"
+  # Inherits temperature and max_tokens from default_model_settings
+
+maestro_default:
+  primary: "gemini/gemini-2.5-pro"
+  fallback: "deepseek/deepseek-r1-0528:free"
+  # Inherits temperature and max_tokens from default_model_settings
+
+objective_generator:
+  primary: "gemini/gemini-2.5-pro"
+  fallback: "deepseek/deepseek-r1-0528:free"
+  # Inherits temperature and max_tokens from default_model_settings
+
+capacitation_generator:
+  primary: "gemini/gemini-2.5-pro"
+  fallback: "deepseek/deepseek-r1-0528:free"
+  # Inherits temperature and max_tokens from default_model_settings
+
+error_analyzer:
+  primary: "gemini/gemini-2.5-pro"
+  fallback: "deepseek/deepseek-chat-v3-0324:free"
+  # Inherits temperature and max_tokens from default_model_settings
+
+# Example of how to override temperature for a specific model if needed:
+# some_other_model:
+#   primary: "some-model"
+#   fallback: "another-model"
+#   temperature: 0.5
+#   max_tokens: 4000
