+++ pyproject.toml
@@ -22,6 +22,8 @@ dependencies = [
     "typer (>=0.16.0,<0.17.0)"
 ]
 
+[tool.poetry]
+package-mode = false
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
