+++ tests/test_agents.py
@@ -4,7 +4,8 @@ import logging
 from unittest.mock import MagicMock, patch
 import requests
 
-from agent.agents import ArchitectAgent, MaestroAgent, parse_json_response
+from agent.agents import ArchitectAgent, MaestroAgent # parse_json_response moved
+from agent.utils.json_parser import parse_json_response # Import from new location
 from agent.utils.llm_client import call_llm_api
 
 # Logger mockado que pode ser passado para as funções do cérebro
@@ -71,7 +72,7 @@ def test_agents_parse_json_response_invalid_json(mock_logger):
     assert "Erro ao decodificar JSON" in error
 
 # --- Testes para ArchitectAgent ---
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_success(mock_call_llm, mock_logger, model_config):
     valid_patches_json_str = json.dumps({
         "analysis": "Análise detalhada aqui.",
@@ -96,7 +97,7 @@ def test_architect_plan_action_success(mock_call_llm, mock_logger, model_config)
     assert "manifesto" in prompt_arg
 
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_llm_error(mock_call_llm, mock_logger, model_config):
     mock_call_llm.return_value = (None, "Erro de API simulado")
     architect = ArchitectAgent(model_config, mock_logger)
@@ -105,7 +106,7 @@ def test_architect_plan_action_llm_error(mock_call_llm, mock_logger, model_confi
     assert plan_data is None
     assert "Erro ao chamar LLM para plano de patches: Erro de API simulado" in error
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_empty_llm_response(mock_call_llm, mock_logger, model_config):
     mock_call_llm.return_value = ("", None) # Resposta vazia
     architect = ArchitectAgent(model_config, mock_logger)
@@ -114,7 +115,7 @@ def test_architect_plan_action_empty_llm_response(mock_call_llm, mock_logger, mo
     assert "Resposta vazia do LLM para plano de patches" in error
 
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_malformed_json(mock_call_llm, mock_logger, model_config):
     mock_call_llm.return_value = ("json { invalido", None)
     architect = ArchitectAgent(model_config, mock_logger)
@@ -123,7 +124,7 @@ def test_architect_plan_action_malformed_json(mock_call_llm, mock_logger, model_
     assert plan_data is None
     assert "Erro ao fazer parse do JSON do plano de patches: Erro ao decodificar JSON" in error # Erro de parse_json_response
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_json_missing_patches_key(mock_call_llm, mock_logger, model_config):
     invalid_json_str = json.dumps({"analysis": "sem patches"})
     mock_call_llm.return_value = (invalid_json_str, None)
@@ -134,7 +135,7 @@ def test_architect_plan_action_json_missing_patches_key(mock_call_llm, mock_logg
     assert "JSON do plano de patches inválido ou não contém a chave 'patches_to_apply' como uma lista." in error
 
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_architect_plan_action_invalid_patch_structure(mock_call_llm, mock_logger, model_config):
     # Patch INSERT sem content
     invalid_patches_json_str = json.dumps({
@@ -149,7 +150,7 @@ def test_architect_plan_action_invalid_patch_structure(mock_call_llm, mock_logge
 
 
 # --- Testes para MaestroAgent ---
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_success(mock_call_llm, mock_logger, model_config):
     maestro_response_json_str = json.dumps({"strategy_key": "APPLY_AND_TEST"})
     mock_call_llm.return_value = (maestro_response_json_str, None)
@@ -169,7 +170,7 @@ def test_maestro_choose_strategy_success(mock_call_llm, mock_logger, model_confi
     prompt_arg = kwargs["prompt"]
     assert json.dumps(action_plan, ensure_ascii=False, indent=2) in prompt_arg # Verificar se o plano de ação está no prompt
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_api_error_then_success(mock_call_llm, mock_logger, model_config):
     maestro_response_model2_json_str = json.dumps({"strategy_key": "MODEL2_WINS"})
     mock_call_llm.return_value = (None, "Erro API no modelo1") # Simulate a single failed call
@@ -184,7 +185,7 @@ def test_maestro_choose_strategy_api_error_then_success(mock_call_llm, mock_logg
     assert "Erro da API" in decision_logs[0]["raw_response"]
     assert mock_call_llm.call_count == 1 # Only one call to the mocked function
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_parsing_error(mock_call_llm, mock_logger, model_config):
     mock_call_llm.return_value = ("json { invalido", None)
     maestro = MaestroAgent(model_config, {"validation_strategies": {}}, mock_logger)
@@ -194,7 +195,7 @@ def test_maestro_choose_strategy_parsing_error(mock_call_llm, mock_logger, model
     assert decision_logs[0]["success"] is False
     assert "Erro ao fazer parse" in decision_logs[0]["raw_response"]
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_json_schema_invalid(mock_call_llm, mock_logger, model_config):
     mock_call_llm.return_value = (json.dumps({"other_key": "val"}), None)
     maestro = MaestroAgent(model_config, {"validation_strategies": {}}, mock_logger)
@@ -203,7 +204,7 @@ def test_maestro_choose_strategy_json_schema_invalid(mock_call_llm, mock_logger,
     assert decision_logs[0]["success"] is False
     assert "JSON com formato inválido ou faltando 'strategy_key'" in decision_logs[0]["raw_response"]
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_capacitation_required(mock_call_llm, mock_logger, model_config):
     maestro_response_json_str = json.dumps({"strategy_key": "CAPACITATION_REQUIRED"})
     mock_call_llm.return_value = (maestro_response_json_str, None)
@@ -213,7 +214,7 @@ def test_maestro_choose_strategy_capacitation_required(mock_call_llm, mock_logge
     assert decision_logs[0]["success"] is True
     assert decision_logs[0]["parsed_json"] == {"strategy_key": "CAPACITATION_REQUIRED"}
 
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_web_search_required(mock_call_llm, mock_logger, model_config):
     """Test that MaestroAgent can return WEB_SEARCH_REQUIRED strategy"""
     maestro_response_json_str = json.dumps({"strategy_key": "WEB_SEARCH_REQUIRED"})
@@ -223,7 +224,7 @@ def test_maestro_choose_strategy_web_search_required(mock_call_llm, mock_logger,
 
     assert decision_logs[0]["success"] is True
     assert decision_logs[0]["parsed_json"] == {"strategy_key": "WEB_SEARCH_REQUIRED"}
-@patch('agent.agents.call_llm_api')
+@patch('agent.utils.llm_client.call_llm_api')
 def test_maestro_choose_strategy_with_memory_summary(mock_call_llm, mock_logger, model_config):
     maestro_response_json_str = json.dumps({"strategy_key": "STRATEGY_WITH_MEMORY"})
     mock_call_llm.return_value = (maestro_response_json_str, None)
